ChatModels/Gemini_ChatModel.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv

load_dotenv()

model = ChatGoogleGenerativeAI(model='gemini-2.5')

result = model.invoke('What is national food of Pakistan?')

print(result.content)