requirements.txt
# Environment 
python-dotenv

# LangChain
langchain
langchain-core

# Chat Models
langchain-openai
openai
langchain-anthropic
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
huggingface-hub

# Other Utilities
numpy
scikit-learn
